Move sub-classes of "DataGridViewComboBoxEditingControl", "ComponentEditorForm" and "DocComment" classes to their own files
